Created the strategy folder to hold the strategy
strategy is a logic to extract data from the token.
To do that I created a valitade method inside the jwt.strategy that
automatically append the data from the token to the req.user field, inside the
request. Then I use the data appended in the request to query the DB for the
user and returning it