created end2end testing for the signup and login endpoints
For testing, Im using Jest and pactum.
Pactum allows me to create it() functions that receive one message
and one arrow function which is going to perform the testing.
The pactum has method to create all sort of requests and expect
result from the endpoint, so I can create many different test cases
and they are going to run and check is the app is working end to end.